Minor Fixes
- Fixed fields that are long types not working with modifiers that use decimals (bluefire having no cooldowns)
- Fixed AvatarState triggering when the damage event is cancelled (#1346)
- Fixed WaterSpoutWave going backwards in AvatarState (#1344)

+++ core/src/com/projectkorra/projectkorra/attribute/AttributeModifier.java
@@ -30,9 +30,9 @@ public enum AttributeModifier {
 		} else if (oldValue instanceof Float) {
 			return oldValue.floatValue() * modifier.floatValue();
 		} else if (oldValue instanceof Long) {
-			return oldValue.longValue() * modifier.longValue();
+			return (long) (oldValue.longValue() * modifier.doubleValue());
 		} else if (oldValue instanceof Integer) {
-			return oldValue.intValue() * modifier.intValue();
+			return (int) (oldValue.intValue() * modifier.doubleValue());
 		}
 		return 0;
 	}), DIVISION((oldValue, modifier) -> {
@@ -41,9 +41,9 @@ public enum AttributeModifier {
 		} else if (oldValue instanceof Float) {
 			return oldValue.floatValue() / modifier.floatValue();
 		} else if (oldValue instanceof Long) {
-			return oldValue.longValue() / modifier.longValue();
+			return (long) (oldValue.longValue() / modifier.doubleValue());
 		} else if (oldValue instanceof Integer) {
-			return oldValue.intValue() / modifier.intValue();
+			return (int) (oldValue.intValue() / modifier.doubleValue());
 		}
 		return 0;
 	}), SET((oldValue, modifier) -> modifier);
